update readme and examples
