+++ examples/ImportC++/jerry_expmple_addCplusplus.cpp
